move tests and create subfolders
