+++ lib/web_ui/test/html/bitmap_canvas_golden_test.dart
@@ -11,7 +11,7 @@ import 'package:ui/src/engine.dart';
 import 'package:ui/ui.dart';
 
 import 'package:web_engine_tester/golden_tester.dart';
-import '../golden_tests/engine/screenshot.dart';
+import 'screenshot.dart';
 
 void main() {
   internalBootstrapBrowserTest(() => testMain);
