+++ lib/web_ui/test/html/canvas_clip_path_test.dart
@@ -10,7 +10,7 @@ import 'package:test/test.dart';
 import 'package:ui/ui.dart' hide TextStyle;
 import 'package:ui/src/engine.dart' as engine;
 
-import '../golden_tests/engine/screenshot.dart';
+import 'screenshot.dart';
 
 void main() {
   internalBootstrapBrowserTest(() => testMain);
