+++ lib/web_ui/test/html/clip_op_golden_test.dart
