+++ lib/web_ui/test/html/compositing/canvas_blend_golden_test.dart
@@ -11,7 +11,7 @@ import 'package:ui/ui.dart' hide TextStyle;
 import 'package:ui/src/engine.dart';
 
 import 'package:web_engine_tester/golden_tester.dart';
-import '../golden_tests/engine/screenshot.dart';
+import '../screenshot.dart';
 
 void main() {
   internalBootstrapBrowserTest(() => testMain);
