+++ lib/web_ui/test/html/compositing/canvas_image_blend_mode_test.dart
@@ -7,8 +7,8 @@ import 'package:test/test.dart';
 import 'package:ui/ui.dart' hide TextStyle;
 import 'package:ui/src/engine.dart';
 
-import '../golden_tests/engine/screenshot.dart';
-import '../golden_tests/engine/testimage.dart';
+import '../screenshot.dart';
+import '../testimage.dart';
 
 void main() {
   internalBootstrapBrowserTest(() => testMain);
