+++ lib/web_ui/test/html/compositing/color_filter_golden_test.dart
