+++ lib/web_ui/test/html/compositing/compositing_golden_test.dart
