+++ lib/web_ui/test/html/drawing/canvas_arc_golden_test.dart
