+++ lib/web_ui/test/html/drawing/canvas_draw_image_golden_test.dart
@@ -13,7 +13,7 @@ import 'package:ui/src/engine.dart';
 
 import 'package:web_engine_tester/golden_tester.dart';
 
-import '../golden_tests/engine/screenshot.dart';
+import '../screenshot.dart';
 
 void main() {
   internalBootstrapBrowserTest(() => testMain);
