+++ lib/web_ui/test/html/drawing/canvas_draw_picture_test.dart
@@ -9,7 +9,7 @@ import 'package:test/test.dart';
 import 'package:ui/ui.dart';
 import 'package:ui/src/engine.dart';
 
-import '../golden_tests/engine/screenshot.dart';
+import '../screenshot.dart';
 
 final Rect region = Rect.fromLTWH(0, 0, 500, 100);
