+++ lib/web_ui/test/html/drawing/canvas_draw_points_test.dart
