+++ lib/web_ui/test/html/drawing/canvas_lines_golden_test.dart
