+++ lib/web_ui/test/html/drawing/canvas_rect_golden_test.dart
