+++ lib/web_ui/test/html/drawing/canvas_rrect_golden_test.dart
