+++ lib/web_ui/test/html/drawing/canvas_stroke_joins_golden_test.dart
