+++ lib/web_ui/test/html/drawing/canvas_stroke_rects_golden_test.dart
