+++ lib/web_ui/test/html/drawing/draw_vertices_golden_test.dart
