+++ lib/web_ui/test/html/paragraph/text_multiline_clipping_golden_test.dart
@@ -8,7 +8,7 @@ import 'package:test/bootstrap/browser.dart';
 import 'package:ui/src/engine.dart';
 import 'package:ui/ui.dart';
 
-import 'scuba.dart';
+import 'text_scuba.dart';
 
 typedef PaintTest = void Function(RecordingCanvas recordingCanvas);
