+++ lib/web_ui/test/html/path_metrics_test.dart
@@ -8,7 +8,7 @@ import 'package:test/bootstrap/browser.dart';
 import 'package:test/test.dart';
 import 'package:ui/ui.dart' hide TextStyle;
 import 'package:ui/src/engine.dart';
-import '../../matchers.dart';
+import '../matchers.dart';
 
 import 'package:web_engine_tester/golden_tester.dart';
