+++ lib/web_ui/test/html/path_to_svg_golden_test.dart
